Moved pl loggers to mttl.logging
