+++ mttl/utils.py
@@ -5,7 +5,6 @@ import random
 import string
 from typing import Optional
 
-import pytorch_lightning as pl
 import torch
 import torch.nn as nn
 from pytorch_lightning.utilities.rank_zero import rank_zero_only
@@ -138,81 +137,6 @@ def agg_dicts(list_of_dicts, agg="mean", tag=False):
         for k, v in out.items():
             out[k] = v / len(list_of_dicts)
     return out
-
-
-def get_pl_loggers(args):
-    loggers = []
-
-    add_simple_logger(loggers, args)
-    add_tb_logger(loggers, args)
-    add_wandb_logger(loggers, args)
-    add_mlf_logger(loggers)
-
-    return loggers
-
-
-def add_simple_logger(loggers, args):
-    from mttl.models.utils import SimpleLogger
-
-    loggers.append(SimpleLogger(args.output_dir))
-
-
-def add_tb_logger(loggers, args):
-    if args.tensorboard:
-        tb_logger = pl.loggers.TensorBoardLogger(save_dir=args.output_dir)
-        loggers.append(tb_logger)
-
-
-def add_wandb_logger(loggers, args) -> pl.loggers.WandbLogger:
-    if not os.environ.get("WANDB_API_KEY"):
-        return
-
-    import wandb
-
-    if not args.exp_name:
-        args.exp_name = os.environ.get("AMLT_JOB_NAME")
-    if not args.wandb_project:
-        args.wandb_project = os.environ.get("WANDB_PROJECT")
-
-    wandb_logger = pl.loggers.WandbLogger(
-        project=args.wandb_project,
-        name=args.exp_name,
-        settings=wandb.Settings(start_method="fork"),
-    )
-    wandb_logger.experiment.save("*.py")
-    loggers.append(wandb_logger)
-
-
-def add_mlf_logger(loggers):
-    import pytorch_lightning as pl
-    from pytorch_lightning.utilities.rank_zero import rank_zero_only
-
-    class MLFlowLoggerCustom(pl.loggers.MLFlowLogger):
-        def __init__(self, *args, **kwargs):
-            super().__init__(*args, **kwargs)
-
-        @rank_zero_only
-        def log_hyperparams(self, *args, **kwargs) -> None:
-            try:
-                super().log_hyperparams(*args, **kwargs)
-            except:
-                pass
-
-    try:
-        from azureml.core.run import Run
-
-        run = Run.get_context()
-
-        mlf_logger = MLFlowLoggerCustom(
-            experiment_name=run.experiment.name,
-        )
-        mlf_logger._run_id = run.id
-    except:
-        logger.warning("Couldn't instantiate MLFlowLogger!")
-        mlf_logger = None
-
-    if mlf_logger is not None:
-        loggers.append(mlf_logger)
 
 
 def get_checkpoint_path(path, step=None, use_last=False):
