+++ projects/modular_llm/finetune_experts.py
@@ -11,7 +11,7 @@ from typing import Callable
 
 from mttl.callbacks import LiveCheckpointCallback, RougeCallback
 from mttl.datamodule.base import get_datamodule
-from mttl.logging import logger, setup_logging
+from mttl.logging import get_pl_loggers, logger, setup_logging
 from mttl.models.expert_config import ExpertConfig
 from mttl.models.expert_model import ExpertModel as ExpertTrainer
 from mttl.models.expert_model import MoEModel as MoETrainer
@@ -32,7 +32,7 @@ from mttl.models.library.library_transforms import (
 )
 from mttl.models.modifiers.base import ModifierConfig
 from mttl.models.monitors import get_monitors
-from mttl.utils import get_checkpoint_path, get_pl_loggers, remote_login
+from mttl.utils import get_checkpoint_path, remote_login
 from projects.modular_llm.src.callbacks import DownstreamEvalCallback
 from projects.modular_llm.src.retrievers import RandomRetriever, SVDEmbeddingRetriever
