+++ projects/modular_llm/train_experts_main.py
@@ -10,18 +10,13 @@ sys.path.append(os.path.join(os.path.dirname(__file__), "..", ".."))
 
 from mttl.callbacks import LiveCheckpointCallback, NanoMMLUCallback, RougeCallback
 from mttl.datamodule.base import get_datamodule
-from mttl.logging import logger, setup_logging
+from mttl.logging import get_pl_loggers, logger, setup_logging
 from mttl.models.expert_config import ExpertConfig
 from mttl.models.expert_model import ExpertModel, MoEModel
 from mttl.models.library.expert import Expert, load_expert
 from mttl.models.library.expert_library import ExpertLibrary, LocalExpertLibrary
 from mttl.models.monitors import get_monitors
-from mttl.utils import (
-    generate_random_string,
-    get_pl_loggers,
-    rank_zero_only_and_wait,
-    remote_login,
-)
+from mttl.utils import generate_random_string, rank_zero_only_and_wait, remote_login
 from projects.modular_llm.src.callbacks import DownstreamEvalCallback
 from projects.modular_llm.src.transfer_matrix import TransferMatrixConfig
 from projects.modular_llm.src.transfer_matrix import run_eval as produce_transfer_matrix
